Move debug camera to separate view
